py/fractribution_model/requirements.txt
google-cloud-bigquery
functions-framework==1.3.0
